Reorganizing defaultdict2
